+++ test/type_inference/init.lua
@@ -547,3 +547,18 @@ local t
 for k, <?v?> in pairs(t) do
 end
 ]]
+
+TEST 'boolean' [[
+---@generic T: table, V
+---@param t T
+---@return fun(table: V[], i?: integer):integer, V
+---@return T
+---@return integer i
+local function ipairs(t) end
+
+---@type boolean[]
+local t
+
+for _, <?v?> in ipairs(t) do
+end
+]]