+++ test/type_inference/init.lua
@@ -87,7 +87,7 @@ TEST 'table' [[
 local <?t?> = {}
 ]]
 
-TEST 'function' [[
+TEST 'any' [[
 <?x?>()
 ]]